refactor: update transform-es2015-modules-commonjs
#2300
