+++ visualization/app/codeCharta/state/selectors/accumulatedData/metricData/nodeMetricData.selector.ts
@@ -3,7 +3,7 @@ import { hierarchy } from "d3-hierarchy"
 import { BlacklistItem, BlacklistType, NodeMetricData } from "../../../../codeCharta.model"
 import { FileState } from "../../../../model/files/files"
 import { isLeaf, isPathBlacklisted } from "../../../../util/codeMapHelper"
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { blacklistSelector } from "../../../store/fileSettings/blacklist/blacklist.selector"
 import { visibleFileStatesSelector } from "../../visibleFileStates.selector"
 import { sortByMetricName } from "./sortByMetricName"
