+++ visualization/app/codeCharta/state/selectors/searchedNodes/searchedNodePaths.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../angular-redux/store"
+import { createSelector } from "../../angular-redux/createSelector"
 import { searchedNodesSelector } from "./searchedNodes.selector"
 
 export const searchedNodePathsSelector = createSelector([searchedNodesSelector], searchedNodes => new Set(searchedNodes.map(x => x.path)))
