+++ visualization/app/codeCharta/state/selectors/selectedNode.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../angular-redux/store"
+import { createSelector } from "../angular-redux/createSelector"
 import { selectedBuildingIdSelector } from "../store/appStatus/selectedBuildingId/selectedBuildingId.selector"
 import { idToNodeSelector } from "./accumulatedData/idToNode.selector"
