+++ visualization/app/codeCharta/state/store/appSettings/isEdgeMetricVisible/isEdgeMetricVisible.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { appSettingsSelector } from "../appSettings.selector"
 
 export const isEdgeMetricVisibleSelector = createSelector([appSettingsSelector], appSettings => appSettings.isEdgeMetricVisible)
