+++ visualization/app/codeCharta/state/store/appSettings/isHeightAndColorMetricLinked/isColorMetricLinkedToHeightMetric.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { appSettingsSelector } from "../appSettings.selector"
 
 export const isColorMetricLinkedToHeightMetricSelector = createSelector(
