+++ visualization/app/codeCharta/state/store/dynamicSettings/focusedNodePath/focusedNodePath.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { dynamicSettingsSelector } from "../dynamicSettings.selector"
 
 export const focusedNodePathSelector = createSelector([dynamicSettingsSelector], dynamicSettings => dynamicSettings.focusedNodePath)
