+++ visualization/app/codeCharta/state/store/dynamicSettings/margin/margin.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { dynamicSettingsSelector } from "../dynamicSettings.selector"
 
 export const marginSelector = createSelector([dynamicSettingsSelector], dynamicSettings => dynamicSettings.margin)
