+++ visualization/app/codeCharta/state/store/fileSettings/markedPackages/markedPackages.selector.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "../../../angular-redux/store"
+import { createSelector } from "../../../angular-redux/createSelector"
 import { fileSettingsSelector } from "../fileSettings.selector"
 
 export const markedPackagesSelector = createSelector([fileSettingsSelector], fileSettings => fileSettings.markedPackages)
