+++ visualization/app/codeCharta/ui/attributeSideBar/util/showDeltaValueSelector.ts
@@ -1,5 +1,5 @@
 import { Node } from "../../../codeCharta.model"
-import { createSelector } from "../../../state/angular-redux/store"
+import { createSelector } from "../../../state/angular-redux/createSelector"
 import { selectedNodeSelector } from "../../../state/selectors/selectedNode.selector"
 
 export const _shouldShowDeltaValue = (node?: Pick<Node, "deltas">) => Boolean(node?.deltas)
