+++ visualization/babel.config.json
@@ -14,7 +14,7 @@
 	],
 	"env": {
 		"test": {
-			"plugins": ["transform-es2015-modules-commonjs"]
+			"plugins": ["@babel/plugin-transform-modules-commonjs"]
 		}
 	}
 }
