+++ visualization/package.json
@@ -106,6 +106,7 @@
 	"devDependencies": {
 		"@babel/cli": "^7.13.14",
 		"@babel/core": "^7.13.15",
+		"@babel/plugin-transform-modules-commonjs": "^7.19.6",
 		"@babel/preset-env": "^7.13.15",
 		"@babel/preset-typescript": "^7.13.0",
 		"@testing-library/angular": "^10.11.0",
@@ -122,7 +123,6 @@
 		"@typescript-eslint/eslint-plugin": "^5.0.0",
 		"@typescript-eslint/parser": "^5.0.0",
 		"babel-loader": "^9.1.0",
-		"babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
 		"bestzip": "^2.2.0",
 		"clean-webpack-plugin": "^4.0.0",
 		"css-loader": "^6.0.0",
